*/Dockerfile.template: mkdir -p /etc/portage/package.use
With the stage3 from 20150402 the directory already exists:

  # cat /etc/portage/package.use/iputils
  net-misc/iputils -caps -filecaps

I don't know how reliable that's going to be though, so I'm not
dropping the mkdir call entirely.

Dockerfile.templates that used to write to an /etc/portage/package.use
file now write to a package-specific file inside the package.use
directory (which is better practice anyway).